fix(auth): move login/register to (redirected) subgroup and fix onboarding URL
- Move login and register pages to (auth)/(redirected)/ sub-route-group
- Empty (auth)/layout.tsx to allow /verify-email access for logged-in users
- Fix onboarding redirect URL from /onboarding/workspace/new to /workspace/new
- Apply lint fixes to VerifyEmailPageClient and useVerifyEmail
