+++ apps/web/src/app/(auth)/(redirected)/register/page.tsx
