+++ apps/web/src/app/(auth)/verify-email/VerifyEmailPageClient.tsx
@@ -1,15 +1,15 @@
 "use client";
 
 import { Loading02Icon, Mail01Icon } from "@hugeicons/core-free-icons";
+import { AuthLayout } from "@/components/auth/AuthLayout";
+import { useVerifyEmail } from "@/components/auth/hook/useVerifyEmail";
+import { Icon } from "@/components/shared/Icon";
+import { Button } from "@/components/ui/button";
 import {
   InputOTP,
   InputOTPGroup,
   InputOTPSlot,
 } from "@/components/ui/input-otp";
-import { Icon } from "@/components/shared/Icon";
-import { Button } from "@/components/ui/button";
-import { AuthLayout } from "@/components/auth/AuthLayout";
-import { useVerifyEmail } from "@/components/auth/hook/useVerifyEmail";
 import { EMAIL_OTP_LENGTH } from "@/lib/email/constants";
 
 interface Props {
@@ -23,26 +23,17 @@ export function VerifyEmailPageClient({
   alreadyVerified,
   callbackURL,
 }: Props) {
-  const {
-    otp,
-    setOtp,
-    isVerifying,
-    isSending,
-    error,
-    cooldown,
-    handleResend,
-  } = useVerifyEmail({ email, alreadyVerified, callbackURL });
+  const { otp, setOtp, isVerifying, isSending, error, cooldown, handleResend } =
+    useVerifyEmail({ email, alreadyVerified, callbackURL });
 
   return (
     <AuthLayout
       title="Confirme seu email"
       description={`Digite o código de ${EMAIL_OTP_LENGTH} dígitos enviado para ${email}`}
       footer={
-        callbackURL ? (
-          <>Após confirmar, você será redirecionado para concluir a ação.</>
-        ) : (
-          <>Não tem mais acesso a este email? Fale com o suporte.</>
-        )
+        callbackURL
+          ? "Após confirmar, você será redirecionado para concluir a ação."
+          : "Não tem mais acesso a este email? Fale com o suporte."
       }
     >
       <div className="flex flex-col gap-6 w-full">
@@ -73,6 +64,7 @@ export function VerifyEmailPageClient({
           >
             <InputOTPGroup>
               {Array.from({ length: EMAIL_OTP_LENGTH }).map((_, i) => (
+                // biome-ignore lint/suspicious/noArrayIndexKey: Index is necessary for a realization of control of input OTP
                 <InputOTPSlot key={i} index={i} />
               ))}
             </InputOTPGroup>
